Extract TERMINFO-PATH
LOAD-TERMINFO was iterating over all the possible locations for a
given terminfo database, attempting to open the file, if the file was
opened successfully then parse the terminfo database and return.

Instead have the extract the searching for the locating of the term
info database to a separate function so that load-terminfo only takes
care of parsing the database.